pyproject.toml
[tool.poetry]
name = "reckonup"
version = "0.1.0"
description = ""
authors = ["Yoshio Miyamae <yoshio.miyamae@famoce-succellion.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
django-filter = "^21.1"
Pillow = "^8.4.0"
django-cors-headers = "^3.10.1"
drf-writable-nested = "^0.6.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
django-admin = "^2.0.1"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
mypy = "^0.910"
django-revproxy = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
